Display correct value for verbose control point cursor for Midi CC (int) and plugin automation (map to parameter range). Use double throughout for automation control point y coordinate instead of casting to this and that losing precision. Fix crash bug in route automation menu.
git-svn-id: svn://localhost/ardour2/trunk@2107 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ gtk2_ardour/automation_line.h
@@ -146,7 +146,7 @@ class AutomationLine : public sigc::trackable, public PBD::StatefulThingWithGoin
 	void show_selection();
 	void hide_selection ();
 
-	virtual string  get_verbose_cursor_string (float);
+	virtual string get_verbose_cursor_string (double);
 	virtual void view_to_model_y (double&);
 	virtual void model_to_view_y (double&);
